docs: Add proxy settings to Markdown build
The initial port of HTML code to Markdown didn't include GRASS_PROXY support for the build. This moves the proxy function out of HTML to the common code and calls that function the Markdown build script.

It also fixes typo in a docstring.
